# fix(ingest-limits): Use stripe locking for metadata