--- a/pkg/limits/http.go
+++ b/pkg/limits/http.go
@@ -19,11 +19,6 @@ type httpTenantLimitsResponse struct {
 // ServeHTTP implements the http.Handler interface.
 // It returns the current stream counts and status per tenant as a JSON response.
 func (s *IngestLimits) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	// TODO(grobinson): Avoid acquiring the mutex for the entire duration
-	// of the request.
-	s.mtx.RLock()
-	defer s.mtx.RUnlock()
-
 	tenant := mux.Vars(r)["tenant"]
 	if tenant == "" {
 		http.Error(w, "invalid tenant", http.StatusBadRequest)
@@ -43,16 +38,14 @@ func (s *IngestLimits) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		response      httpTenantLimitsResponse
 	)
 
-	for _, partitions := range s.metadata[tenant] {
-		for _, stream := range partitions {
-			if stream.lastSeenAt >= cutoff {
-				activeStreams++
+	for stream := range s.streams(tenant) {
+		if stream.lastSeenAt >= cutoff {
+			activeStreams++
 
-				// Calculate size only within the rate window
-				for _, bucket := range stream.rateBuckets {
-					if bucket.timestamp >= rateWindowCutoff {
-						totalSize += bucket.size
-					}
+			// Calculate size only within the rate window
+			for _, bucket := range stream.rateBuckets {
+				if bucket.timestamp >= rateWindowCutoff {
+					totalSize += bucket.size
 				}
 			}
 		}