--- a/pkg/limits/http.go
+++ b/pkg/limits/http.go
@@ -38,7 +38,7 @@ func (s *IngestLimits) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		response      httpTenantLimitsResponse
 	)
 
-	s.metadata.All(func(tenantID string, partitionID int32, stream Stream) {
+	s.metadata.All(func(tenantID string, _ int32, stream Stream) {
 		if tenantID != tenant {
 			return
 		}