**What this PR does / why we need it**:
This pull request implements a stripe-lock approach to read/write the internal ingest-limits stream metadata structure. The stripes are distributed over the total amount of partitions (e.g. default is 64). This allows updating and reading the structure with less lock contention. 

```
                                                                                    │    old.txt    │                new.txt                │
                                                                                   │    sec/op     │    sec/op     vs base                 │
StreamMetadata_Store/4_partitions_small_streams_single_tenant_create-14               180.8n ± ∞ ¹   133.7n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/4_partitions_small_streams_single_tenant_update-14               190.6n ± ∞ ¹   135.5n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/4_partitions_small_streams_single_tenant_create_parallel-14      320.4n ± ∞ ¹   232.3n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/4_partitions_small_streams_single_tenant_update_parallel-14      347.7n ± ∞ ¹   256.7n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_create-14               189.1n ± ∞ ¹   124.0n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_update-14               195.1n ± ∞ ¹   125.9n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_create_parallel-14      341.6n ± ∞ ¹   125.1n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_update_parallel-14      356.0n ± ∞ ¹   150.3n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_create-14               206.9n ± ∞ ¹   134.7n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_update-14               208.6n ± ∞ ¹   136.6n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_create_parallel-14     336.20n ± ∞ ¹   82.15n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_update_parallel-14     343.20n ± ∞ ¹   91.31n ± ∞ ¹        ~ (p=1.000 n=1) ²
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_create-14              211.4n ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_update-14              212.2n ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_create_parallel-14     327.2n ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_update_parallel-14     332.0n ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__create-14                            134.9n ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__update-14                            139.1n ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__create_parallel-14                   60.90n ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__update_parallel-14                   65.67n ± ∞ ¹
geomean                                                                               259.3n         124.4n        -46.83%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ benchmark set differs from baseline; geomeans may not be comparable

                                                                                   │   old.txt   │               new.txt                │
                                                                                   │    B/op     │     B/op      vs base                │
StreamMetadata_Store/4_partitions_small_streams_single_tenant_create-14              56.00 ± ∞ ¹    56.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/4_partitions_small_streams_single_tenant_update-14              72.00 ± ∞ ¹    71.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
StreamMetadata_Store/4_partitions_small_streams_single_tenant_create_parallel-14     90.00 ± ∞ ¹    90.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/4_partitions_small_streams_single_tenant_update_parallel-14     157.0 ± ∞ ¹    141.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_create-14              56.00 ± ∞ ¹    56.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_update-14              72.00 ± ∞ ¹    72.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_create_parallel-14     89.00 ± ∞ ¹    90.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_update_parallel-14     137.0 ± ∞ ¹    138.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_create-14              56.00 ± ∞ ¹    56.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_update-14              72.00 ± ∞ ¹    72.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_create_parallel-14     73.00 ± ∞ ¹   116.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_update_parallel-14     113.0 ± ∞ ¹    159.0 ± ∞ ¹       ~ (p=1.000 n=1) ³
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_create-14             56.00 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_update-14             72.00 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_create_parallel-14    62.00 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_update_parallel-14    92.00 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__create-14                           56.00 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__update-14                           71.00 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__create_parallel-14                  95.00 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__update_parallel-14                  138.0 ± ∞ ¹
geomean                                                                              78.65          86.48        +6.02%               ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ need >= 4 samples to detect a difference at alpha level 0.05
⁴ benchmark set differs from baseline; geomeans may not be comparable

                                                                                   │   old.txt   │               new.txt               │
                                                                                   │  allocs/op  │  allocs/op   vs base                │
StreamMetadata_Store/4_partitions_small_streams_single_tenant_create-14              2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/4_partitions_small_streams_single_tenant_update-14              2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/4_partitions_small_streams_single_tenant_create_parallel-14     2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/4_partitions_small_streams_single_tenant_update_parallel-14     2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_create-14              2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_update-14              2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_create_parallel-14     2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/8_partitions_medium_streams_multi_tenant_update_parallel-14     2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_create-14              2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_update-14              2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_create_parallel-14     2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/16_partitions_large_streams_multi_tenant_update_parallel-14     2.000 ± ∞ ¹   2.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_create-14             2.000 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_update-14             2.000 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_create_parallel-14    2.000 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant_update_parallel-14    2.000 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__create-14                          2.000 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__update-14                          2.000 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__create_parallel-14                 2.000 ± ∞ ¹
StreamMetadata_Store/32_partitions_xlarge_streams_multi_tenant__update_parallel-14                 2.000 ± ∞ ¹
geomean                                                                              2.000         2.000        +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ benchmark set differs from baseline; geomeans may not be comparable
```

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Special notes for your reviewer**:

**Checklist**
- [x] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- [ ] Documentation added
- [x] Tests updated
- [x] Title matches the required conventional commits format, see [here](https://www.conventionalcommits.org/en/v1.0.0/)
  - **Note** that Promtail is considered to be feature complete, and future development for logs collection will be in [Grafana Alloy](https://github.com/grafana/alloy). As such, `feat` PRs are unlikely to be accepted unless a case can be made for the feature actually being a bug fix to existing behavior.
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/setup/upgrade/_index.md`
- [ ] If the change is deprecating or removing a configuration option, update the `deprecated-config.yaml` and `deleted-config.yaml` files respectively in the `tools/deprecated-config-checker` directory. [Example PR](https://github.com/grafana/loki/pull/10840/commits/0d4416a4b03739583349934b96f272fb4f685d15)